Adding Index Hint
Sometimes we know the filter involved in the query and it will be good if we could help OJAI a little bit by given a hint about the know index. If we give an index hint, only the first reader will add it to the queries this is for avoiding duplicate results. The problem is that once given an index hint, we cannot control from where we are reading, so readers can overlap, to avoid this, we add the hints only to the first reader.