Adding Concurrent Context Support
The idea is to be able able to control how many concurrent OJAI queries are spawn at a time. Running sequentially is a really bad idea, but it is worst to create to many thread to run concurrently, specially if we have an unbounded number of concurrent queries where we have to wait for at the same time.

Improving this means execute many current OJAI queries within a bounded ThreadPool and waiting for the results using a sliding window which is better than wait for all at once.